@@ -13,15 +13,19 @@
     "chess.js": "^0.10.3",
     "firebase": "^8.3.1",
     "firebase-tools": "^9.23.0",
+    "ipfs": "^0.60.2",
     "konva": "6.0.0",
     "mdbreact": "^5.2.0",
-    "react": "^16.13.1",
+    "orbit-db": "^0.28.0",
+    "react": "17.0.2",
     "react-component-countdown-timer": "^0.1.8",
     "react-cookie": "^4.1.1",
-    "react-dom": "^16.13.1",
+    "react-dom": "^17.0.2",
     "react-konva": "^16.13.0-3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
+    "react-router-dom": "^6.1.1",
+    "react-scripts": "4.0.3",
+    "semantic-ui-css": "^2.4.1",
+    "semantic-ui-react": "^2.0.4",
     "simple-peer": "^9.7.2",
     "socket.io-client": "2.3.0",
     "styled-components": "^5.1.1",
@@ -30,11 +34,7 @@
     "use-sound": "^1.0.2",
     "uuid": "^7.0.3",
     "web-vitals": "^1.1.2",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.0.4",
-    "web3": "^1.6.1",
-    "ipfs": "^0.60.2",
-    "orbit-db": "^0.28.0"
+    "web3": "^1.6.1"
   },
   "scripts": {
     "start": "react-scripts start",
